NativeFeatureParser extracted from NativeTestCase
